Refactor snippy.lua
