+++ test/unit/reader_spec.lua
@@ -1,6 +1,10 @@
 local snippy = require('snippy')
 
 describe('Snippet reader', function()
+    setup(function()
+        vim.cmd([[runtime! plugin/snippy.vim]])
+    end)
+
     before_each(function()
         snippy.clear_cache()
     end)
